% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-funs.R
\name{to_element_id}
\alias{to_element_id}
\title{Map a TCM variable name to its CLIDATA element abbreviation}
\usage{
to_element_id(x)
}
\arguments{
\item{x}{\code{character}. Internal TCM variable name(s) (e.g. \code{"TMIN"},
\code{"TS-10"}, \code{"WDD"}).}
}
\value{
\code{character} vector of CLIDATA element abbreviations
(\code{eg_el_abbreviation}). Returns \code{NA} with a warning for unrecognised
names.
}
\description{
Map a TCM variable name to its CLIDATA element abbreviation
}
\examples{
to_element_id("Tn")    # "TMIN"
to_element_id("T-10")   # "TS-10"
to_element_id("UNKNOWN") # NA + warning

}
