% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\name{get_sms_max_age}
\alias{get_sms_max_age}
\title{Get the maximum SMS age}
\usage{
get_sms_max_age()
}
\value{
An \code{integer} scalar.
}
\description{
Returns the current maximum acceptable age (in days) for SMS messages,
as set by \code{SMSCOLLECTR_MAX_AGE} or \code{set_sms_max_age()}.
}
\examples{
get_sms_max_age()   # 3 (default)

}
\seealso{
\code{\link[=set_sms_max_age]{set_sms_max_age()}}
}
