% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smscollectr-package.R
\docType{package}
\name{smscollectr-package}
\alias{smscollectr}
\alias{smscollectr-package}
\title{smscollectr: Collect and Process Rain Gauge SMS Data}
\description{
Ingests, parses, and exports rain gauge SMS messages from Google Sheets. Supports incremental reading, CLICOM output, and FTP delivery.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oousmane/smscollectr}
  \item Report bugs at \url{https://github.com/oousmane/smscollectr/issues}
}

}
\author{
\strong{Maintainer}: Ousmane Ouedraogo \email{ousmane@anam.bf}

}
\keyword{internal}
