% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smscollectr-package.R
\docType{package}
\name{smscollectr-package}
\alias{smscollectr}
\alias{smscollectr-package}
\title{smscollectr: Collect and Process Rain Gauge SMS Data}
\description{
Reads structured SMS messages forwarded to a Google Sheet by field observers, detects and fixes malformed entries, parses valid messages into tidy rainfall and agrometeorological observations, and exports results in a format compatible with the CLIDATA climatological database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oousmane/smscollectr}
  \item \url{https://oousmane.github.io/smscollectr/}
  \item Report bugs at \url{https://github.com/oousmane/smscollectr/issues}
}

}
\author{
\strong{Maintainer}: Ousmane Ouedraogo \email{ouedraogo.ousmane@meteoburkina.bf}

}
\keyword{internal}
